requirements.txt
numpy~=2.3.2
torch~=2.8.0
torchvision~=0.23.0
pillow~=11.3.0
streamlit~=1.49.1